feat: implement Update functionality for EntityType with corresponding tests
