+++ src/Dynamic.Application/Ports/In/EntityTypeQuery/EntityTypeQueryHandler.cs
@@ -21,14 +21,6 @@ public class EntityTypeQueryHandler : IEntityTypeQueryUseCase
     {
         return await _entityTypeRepository.GetByIdAsync(id);
     }
-    public async Task<EntityType?> UpdateAsync(string id, string? name)
-    {
-        var entity = await _entityTypeRepository.GetByIdAsync(id);
-        if (entity == null)
-            return null;
-        entity.Name = name;
-        return await _entityTypeRepository.UpdateAsync(entity);
-    }
     public async Task<bool> DeleteAsync(string id)
     {
         return await _entityTypeRepository.DeleteAsync(id);
