+++ src/Dynamic.Application/Ports/In/UpdateEntityType/IUpdateEntityTypeUseCase.cs
@@ -1,0 +1,8 @@
+using Dynamic.Domain.Models;
+
+namespace Dynamic.Application.Ports.In.UpdateEntityType;
+
+public interface IUpdateEntityTypeUseCase
+{
+    Task<EntityType?> UpdateAsync(string id, string? name);
+}
